# enhance: Custom bitset and bitsetview prototypes