--- a/internal/core/src/bitset/readme.txt
+++ b/internal/core/src/bitset/readme.txt
@@ -1,0 +1,1 @@
+The standlaone version of the bitset library is available at https://github.com/alexanderguzhva/bitset