… internal/core/src/CMakeLists.txt
… internal/core/src/bitset/CMakeLists.txt
… internal/core/src/bitset/bitset.h
… internal/core/src/bitset/detail/element_vectorized.h
… internal/core/src/bitset/detail/element_wise.h
… internal/core/src/bitset/detail/platform/arm/neon-impl.h
… internal/core/src/bitset/detail/platform/arm/neon-inst.cpp
… internal/core/src/bitset/detail/platform/arm/neon.h
… internal/core/src/bitset/detail/platform/arm/sve-decl.h
… internal/core/src/bitset/detail/platform/arm/sve-impl.h
… internal/core/src/bitset/detail/platform/arm/sve-inst.cpp
… internal/core/src/bitset/detail/platform/arm/sve.h
… internal/core/src/bitset/detail/platform/dynamic.cpp
… internal/core/src/bitset/detail/platform/dynamic.h
… internal/core/src/bitset/detail/platform/vectorized_ref.h
… internal/core/src/bitset/detail/platform/x86/avx2-decl.h
… internal/core/src/bitset/detail/platform/x86/avx2-impl.h
… internal/core/src/bitset/detail/platform/x86/avx2-inst.cpp
… internal/core/src/bitset/detail/platform/x86/avx2.h
… internal/core/src/bitset/detail/platform/x86/avx512-decl.h
… internal/core/src/bitset/detail/platform/x86/avx512-impl.h
… internal/core/src/bitset/detail/platform/x86/avx512-inst.cpp
… internal/core/src/bitset/detail/platform/x86/avx512.h
… internal/core/src/bitset/detail/platform/x86/common.h
… internal/core/src/bitset/detail/platform/x86/instruction_set.cpp
… internal/core/src/bitset/detail/platform/x86/instruction_set.h
… internal/core/src/bitset/detail/popcount.h
… internal/core/src/bitset/detail/proxy.h
… internal/core/src/common/CMakeLists.txt
… internal/core/src/common/CustomBitset.h