@liliu-z 2024-02-06
It looks like a right place to optimize
/assign @congqixia for further check