@liliu-z 2024-02-19
Did you try cases in VDBBench with filtering? Curios about the potential improvement