@alexanderguzhva 2024-02-28
/run-cpu-e2e