@alexanderguzhva 2024-03-06
/run-cpu-e2e