@alexanderguzhva 2024-03-07
/run-cpu-e2e