@zhagnlu 2024-03-20
I do exprs tests in my machine, and compare performance between latest and this bitset_simd patch, this pr has better performance for some situation as shows:
![image](https://github.com/milvus-io/milvus/assets/11935707/7ca6aff0-43c8-43a9-8def-5dbedee278a0)