@zhagnlu 2024-03-21
> @zhagnlu what about `Expr.TestMultiLogicalExprsOptimization` results?

It's a test that test conjunct expr optimization, occasional failures may occur due to CPU cache effects. I will update this test  in the future, you can ignore it now.